+++ implementation/js/src/infrastructure/core/remote-data.js
@@ -7,10 +7,12 @@
  * a. url string
  * or 
  * b. object:
- * 	1. + _entity[_id][_method] - string
- *  2. + params(alias:querys) - object
- *  3. + payload - object (payload._id overrides _id)
- *  4. $.ajax options (without -data, -type, -processData, -contentType)
+ *  1. + params(alias:querys) - object
+ *  2. + payload - object (payload.id determines non GET call types)
+ *  3. $.ajax options (without -data, -type, -processData, -contentType)
+ *  	- headers - object (custom http headers)
+ *  	- async - boolean
+ *  	- success/error/complete - fn (you should be using .done/fail/always() after .remote())
  *
  *  Global CROSSDOMAIN Settings - *Deprecated*: set this in a per-request base or use server side proxy
  *  see MDN - https://developer.mozilla.org/en-US/docs/HTTP/Access_control_CORS
@@ -20,10 +22,8 @@
  *  	protocol: '', //https or not? default: '' -> http
  *   	host: '127.0.0.1', 
  *   	port: '5000',
- *   	headers: {
- *   		'Credential': 'user:pwd'/'token',
- *   		...
- *  }
+ *  } 
+ *  Put additional headers for xdomain just in options.headers (same lvl as options.xdomain).
  *  Again, it is always better to use server side proxy/forwarding instead of client side x-domain.
  *
  * events:
@@ -40,6 +40,7 @@
  * 
  * @author Tim Lauv
  * @created 2014.03.24
+ * @updated 2017.07.20
  */ 
 
 ;(function(app, _, $){
@@ -66,22 +67,10 @@
 			//-------------------------------------------------------------------------------------------
 		});
 
-		//process _entity[_id][_method] and strip off options.querys(alias:params)
-		if(options.entity || options._entity){
-			var entity = options.entity || options._entity;
-			options.url = entity;
-		}
-		if(options.payload && options.payload._id){
-			options._id = options.payload._id;
-		}
-		if(options._id || options._method){
-			var url = app.uri(options.url);
-			options.url = url.path(_.compact([url.path(), options._id, options._method]).join('/')).toString();
-		}
+		//fix url?query params (merge with alias querys, +payload.id as ?id=)
 		options.params = _.extend(options.params || {}, options.querys);
-		if(options.params){
-			options.url = (app.uri(options.url)).search(options.params).toString();
-		}
+		if(options.payload && options.payload.id) options.params.id = options.payload.id;
+		options.url = (app.uri(options.url)).search(options.params).toString();
 
 		//app.config.baseAjaxURI
 		if(app.config.baseAjaxURI)
@@ -95,40 +84,61 @@
 			options.xhrFields = _.extend(options.xhrFields || {}, {
 				withCredentials: true //persists session cookies.
 			});
-			options.headers = _.extend(options.headers || {}, crossdomain.headers);
 			// Using another way of setting withCredentials flag to skip FF error in sycned CORS ajax - no cookies tho...:(
 			// options.beforeSend = function(xhr) {
 			// 	xhr.withCredentials = true;
 			// };
 		}
 
+		//+ header: csrf token (standard: Django, for ajax calls to pass through since they already had same origin sandbox check in client browser) 
 		//get csrftoken value from cookie and set to header. (don't forget to set for $.fileupload editor x2 as well)
 		options.headers = options.headers || {};
 		if(app.config.csrftoken && !options.headers[app.config.csrftoken.header])
 			options.headers[app.config.csrftoken.header] = app.cookie.get(app.config.csrftoken.cookie) || 'NOTOKEN';
+
+		//+ header: jwt token (standard: https://jwt.io/introduction)
+		//honor app.config.jwttoken.value as is.
+		if(app.config.jwttoken && !options.headers[app.config.jwttoken.header]){
+			options.headers[app.config.jwttoken.header] = app.config.jwttoken.schema + ' ' + (app.config.jwttoken.value || 'NOTOKEN');
+		}
 
 		app.trigger('app:ajax', options);		
 		return options;
 	}
 
 	_.extend(definition, {
+		//MOCK check
+		mock: function(options){
+			//intercept by ?mock=true in app
+			if(app.param('mock') !== 'true') return;
+
+			var schema = app._mockSchema[options.url.split('?')[0]];
+			if(schema){ //url lvl schemas
+				schema = schema[options.type] || schema['*']; //http method lvl schema record
+				return $.Deferred().resolve(app.mock(schema.schema, schema.provider));
+			}
+		},
 
 		//GET
 		get: function(options, restOpt){
 			options = fixOptions(options, restOpt);
 			options.type = 'GET';
+
 			app.trigger('app:remote-pre-get', options);
-			return $.ajax(options);
+			return this.mock(options) || $.ajax(options);
 		},
 
-		//POST(no payload._id)/PUT/DELETE(payload = {_id: ...})
+		//POST(no payload.id)
+		//PUT(payload = {id: , ...}
+		//DELETE(payload = {id:})
 		change: function(options, restOpt){
 			options = fixOptions(options, restOpt);
 			if(!options.payload) throw new Error('DEV::Core.Remote::payload empty, please use GET');
-			if(options.payload._id && _.size(options.payload) === 1) options.type = 'DELETE';
+
+			if(options.payload.id && _.size(options.payload) === 1) options.type = 'DELETE';
 			else {
 				if(!_.isObject(options.payload)) options.payload = { payload: options.payload };
-				if(!options.payload._id) options.type = 'POST';
+				if(!options.payload.id) options.type = 'POST';
 				else options.type = 'PUT';
 			}
 
@@ -138,7 +148,7 @@
 			}
 
 			app.trigger('app:remote-pre-change', options);
-			return $.ajax(options);
+			return this.mock(options) || $.ajax(options);
 		}
 
 	});